refactor: лайки, удаление, добавление карточек и изменения аватара и информации о себе

+++ src/utils/api.js
@@ -1,8 +1,8 @@
-export class Api {
+class Api {
     constructor({ url, headers }) {
         this._url = url;
         this._headers = headers;
-    }
+    };
 
     _checkResponse(res) {
         if (res.ok) {
@@ -10,7 +10,6 @@ export class Api {
         }
         return Promise.reject(`Ошибка: ${res.status}`);
     };
-
     getInitialCards() {
         return fetch(`${this._url}/cards`, {
             method: 'GET',
@@ -18,7 +17,6 @@ export class Api {
         })
             .then((res) => this._checkResponse(res));
     };
-
     getUserInfo() {
         return fetch(`${this._url}/users/me`, {
             method: 'GET',
@@ -26,7 +24,6 @@ export class Api {
         })
             .then((res) => this._checkResponse(res))
     };
-
     editUserInfo(data) {
         return fetch(`${this._url}/users/me`, {
             method: 'PATCH',
@@ -38,7 +35,16 @@ export class Api {
         })
             .then((res) => this._checkResponse(res));
     };
-
+    editAvatar(data) {
+        return fetch(`${this._url}/users/me/avatar`, {
+            method: 'PATCH',
+            headers: this._headers,
+            body: JSON.stringify({
+                avatar: data.avatar,
+            }),
+        })
+            .then((res) => this._checkResponse(res));
+    };
     addCard(data) {
         return fetch(`${this._url}/cards`, {
             method: 'POST',
@@ -51,7 +57,6 @@ export class Api {
         })
             .then((res) => this._checkResponse(res));
     };
-
     deleteCard(cardId) {
         return fetch(`${this._url}/cards/${cardId}`, {
             method: 'DELETE',
@@ -60,34 +65,20 @@ export class Api {
         })
             .then((res) => this._checkResponse(res));
     };
-
-    likeCard(cardId) {
-        return fetch(`${this._url}/cards/${cardId}/likes`, {
-            method: 'PUT',
-            headers: this._headers,
-            'Content-Type': 'application/json',
-        })
-            .then((res) => this._checkResponse(res));
-    };
-
-    dislike(cardId) {
-        return fetch(`${this._url}/cards/likes/${cardId}`, {
-            method: 'DELETE',
-            headers: this._headers,
-            'Content-Type': 'application/json',
-        })
-            .then((res) => this._checkResponse(res));
-    };
-
-    editAvatar(data) {
-        return fetch(`${this._url}/users/me/avatar`, {
-            method: 'PATCH',
-            headers: this._headers,
-            body: JSON.stringify({
-                avatar: data.avatar,
-            }),
-        })
-            .then((res) => this._checkResponse(res));
+    changeLikeCardStatus(cardId, isLiked) {
+        if (isLiked) {
+            return fetch(`${this._url}/cards/${cardId}/likes`, {
+                method: "PUT",
+                headers: this._headers,
+            })
+                .then((res) => this._checkResponse(res));
+        } else {
+            return fetch(`${this._url}/cards/${cardId}/likes`, {
+                method: 'DELETE',
+                headers: this._headers,
+            })
+                .then((res) => this._checkResponse(res));
+        };
     };
 };
 
@@ -95,7 +86,7 @@ const api = new Api({
     url: 'https://mesto.nomoreparties.co/v1/cohort-62',
     headers: {
         Authorization: 'f71e956f-38a7-4a42-a0a9-eb7efb8f8d45',
-        'Content-Type': 'application/json'
+        'Content-Type': 'application/json',
     }
 });
